remove legacy invitation and anonids from the builder
