+++ openreview/conference/invitation.py
@@ -37,7 +37,6 @@ class SubmissionInvitation(openreview.Invitation):
                     # Only supported for public reviews
                     if submission_stage.create_review_invitation:
                         file_content = file_content.replace("OFFICIAL_REVIEW_NAME = ''", "OFFICIAL_REVIEW_NAME = '" + conference.review_stage.name + "'")
-                    if not conference.legacy_anonids:
                         file_content = file_content.replace("ANON_IDS = False", "ANON_IDS = True")
                         file_content = file_content.replace("DEANONYMIZERS = []", "DEANONYMIZERS = " + json.dumps(conference.get_reviewer_identity_readers('{number}')))
